+++ mecha/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.94.2"
+__version__ = "0.95.2"
 
 
 from .api import *